feat(rust): heavy kafka refactoring, moved portal interceptor from `api` to `tcp` crate
