+++ implementations/rust/ockam/ockam_api/Cargo.toml
@@ -36,7 +36,6 @@ std = [
   "ockam_node/std",
   "ockam_vault/std",
   "ockam_vault_aws/std",
-  "tinyvec/std",
   "tracing/std",
   "storage",
 ]
@@ -95,7 +94,6 @@ sysinfo = "0.30"
 thiserror = "1.0"
 time = { version = "0.3.36", default-features = false, features = ["std", "formatting", "local-offset", "macros"] }
 tiny_http = "0.12.0"
-tinyvec = { version = "1.8.0", features = ["rustc_1_57"] }
 tokio = { version = "1.39.2", features = ["full"] }
 tokio-retry = "0.3.0"
 tracing = { version = "0.1", default-features = false }
