+++ implementations/rust/ockam/ockam_api/src/kafka/inlet_controller.rs
@@ -1,15 +1,11 @@
-use core::str::FromStr;
-use minicbor::Decoder;
+use std::fmt::Debug;
 
 use crate::kafka::kafka_outlet_address;
-use crate::nodes::models::portal::{CreateInlet, InletStatus};
-use crate::nodes::NODEMANAGER_ADDR;
+use crate::nodes::NodeManager;
 use crate::port_range::PortRange;
 use ockam::compat::tokio::sync::Mutex;
 use ockam_abac::PolicyExpression;
-use ockam_core::api::{Request, ResponseHeader, Status};
 use ockam_core::compat::collections::HashMap;
-use ockam_core::compat::net::SocketAddr;
 use ockam_core::compat::rand::random_string;
 use ockam_core::compat::sync::Arc;
 use ockam_core::errcode::{Kind, Origin};
@@ -24,10 +20,21 @@ type BrokerId = i32;
 /// Shared structure for every kafka worker (consumer or producer services)
 /// to keep track of which brokers are being proxied with the relative inlet listener socket address.
 /// Also takes care of creating inlets dynamically when they are not present yet.
-#[derive(Debug, Clone)]
+#[derive(Clone)]
 pub(crate) struct KafkaInletController {
     inner: Arc<Mutex<KafkaInletMapInner>>,
     policy_expression: Option<PolicyExpression>,
+    node_manager: Arc<NodeManager>,
+}
+
+impl Debug for KafkaInletController {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        write!(
+            f,
+            "KafkaInletController {{ policy_expression: {:?}, inlets: {:?} }}",
+            self.policy_expression, self.inner
+        )
+    }
 }
 
 #[derive(Debug)]
@@ -43,6 +50,7 @@ struct KafkaInletMapInner {
 
 impl KafkaInletController {
     pub(crate) fn new(
+        node_manager: Arc<NodeManager>,
         outlet_node_multiaddr: MultiAddr,
         local_interceptor_route: Route,
         remote_interceptor_route: Route,
@@ -60,6 +68,7 @@ impl KafkaInletController {
                 local_interceptor_route,
                 remote_interceptor_route,
             })),
+            node_manager,
             policy_expression,
         }
     }
@@ -93,18 +102,27 @@ impl KafkaInletController {
 
             let inlet_bind_address =
                 HostnamePort::new(inner.bind_hostname.clone(), inner.current_port);
-            Self::request_inlet_creation(
-                context,
-                inlet_bind_address.clone(),
-                inner.outlet_node_multiaddr.clone(),
-                inner.local_interceptor_route.clone(),
-                route![
-                    inner.remote_interceptor_route.clone(),
-                    kafka_outlet_address(broker_id)
-                ],
-                self.policy_expression.clone(),
-            )
-            .await?;
+
+            self.node_manager
+                .create_inlet(
+                    context,
+                    inlet_bind_address.clone(),
+                    inner.local_interceptor_route.clone(),
+                    route![
+                        inner.remote_interceptor_route.clone(),
+                        kafka_outlet_address(broker_id)
+                    ],
+                    inner.outlet_node_multiaddr.clone(),
+                    format!("kafka-inlet-{}", random_string()),
+                    self.policy_expression.clone(),
+                    None,
+                    None,
+                    false,
+                    None,
+                    false,
+                    false,
+                )
+                .await?;
 
             inner.current_port += 1;
             inner
@@ -112,59 +130,6 @@ impl KafkaInletController {
                 .insert(broker_id, inlet_bind_address.clone());
 
             Ok(inlet_bind_address)
-        }
-    }
-
-    async fn request_inlet_creation(
-        context: &Context,
-        inlet_bind_address: HostnamePort,
-        to: MultiAddr,
-        prefix: Route,
-        suffix: Route,
-        policy_expression: Option<PolicyExpression>,
-    ) -> Result<SocketAddr> {
-        let mut payload = CreateInlet::to_node(
-            inlet_bind_address,
-            to,
-            format!("kafka-inlet-{}", random_string()),
-            prefix,
-            suffix,
-            None,
-            true,
-            false,
-            false,
-        );
-        if let Some(expr) = policy_expression {
-            payload.set_policy_expression(expr);
-        }
-        let buffer: Vec<u8> = context
-            .send_and_receive(
-                route![NODEMANAGER_ADDR],
-                Request::post("/node/inlet").body(payload).to_vec()?,
-            )
-            .await?;
-
-        let mut decoder = Decoder::new(&buffer);
-        let response: ResponseHeader = decoder.decode()?;
-
-        let status = response.status().unwrap_or(Status::InternalServerError);
-        if status != Status::Ok {
-            return Err(Error::new(
-                Origin::Transport,
-                Kind::Invalid,
-                format!("cannot create inlet: {}", status),
-            ));
-        }
-        if !response.has_body() {
-            Err(Error::new(
-                Origin::Transport,
-                Kind::Unknown,
-                "invalid create inlet response",
-            ))
-        } else {
-            let status: InletStatus = decoder.decode()?;
-            Ok(SocketAddr::from_str(&status.bind_addr)
-                .map_err(|err| Error::new(Origin::Transport, Kind::Invalid, err))?)
         }
     }
 }
