+++ implementations/rust/ockam/ockam_api/src/kafka/mod.rs
@@ -2,24 +2,19 @@
 //! to the kafka consumer without any modification in the existing application.
 
 mod inlet_controller;
-mod integration_test;
-mod length_delimited;
+pub(crate) mod key_exchange;
 mod outlet_controller;
-mod outlet_service;
-mod portal_listener;
-mod portal_worker;
-mod protocol_aware;
-pub(crate) mod secure_channel_map;
+pub(crate) mod protocol_aware;
+#[cfg(test)]
+mod tests;
 
 pub(crate) use inlet_controller::KafkaInletController;
+pub use key_exchange::{ConsumerPublishing, ConsumerResolution};
 use ockam::identity::Identifier;
 use ockam_abac::expr::{eq, or, str};
 use ockam_abac::{subject_has_credential_policy_expression, subject_identifier_attribute, Expr};
 use ockam_core::Address;
-pub(crate) use outlet_service::OutletManagerService;
-pub(crate) use portal_listener::KafkaPortalListener;
-pub use secure_channel_map::ConsumerPublishing;
-pub use secure_channel_map::ConsumerResolution;
+pub(crate) use outlet_controller::KafkaOutletController;
 
 pub const KAFKA_OUTLET_INTERCEPTOR_ADDRESS: &str = "kafka_interceptor";
 pub const KAFKA_OUTLET_BOOTSTRAP_ADDRESS: &str = "kafka_bootstrap";
