--- implementations/rust/ockam/ockam_api/src/kafka/outlet_service/mod.rs
@@ -1,2 +1,0 @@
-mod interceptor_listener;
-pub(crate) use interceptor_listener::OutletManagerService;
