+++ implementations/rust/ockam/ockam_api/src/kafka/protocol_aware/length_delimited.rs
