+++ implementations/rust/ockam/ockam_api/src/kafka/protocol_aware/mod.rs
@@ -1,6 +1,3 @@
-use crate::kafka::portal_worker::InterceptError;
-use crate::kafka::secure_channel_map::controller::KafkaSecureChannelControllerImpl;
-use crate::kafka::KafkaInletController;
 use bytes::BytesMut;
 use kafka_protocol::messages::ApiKey;
 use minicbor::{CborLen, Decode, Encode};
@@ -9,16 +6,19 @@ use ockam_core::compat::{
     fmt::Debug,
     sync::{Arc, Mutex},
 };
-use ockam_core::{async_trait, Address};
+use ockam_core::{async_trait, Address, Error};
 use ockam_node::Context;
 
-mod metadata_interceptor;
-mod request;
-mod response;
+pub(crate) mod outlet;
 mod tests;
 
+pub(crate) mod inlet;
+mod length_delimited;
 pub(super) mod utils;
-pub(crate) use metadata_interceptor::OutletInterceptorImpl;
+
+use crate::kafka::protocol_aware::length_delimited::{length_encode, KafkaMessageDecoder};
+use ockam_core::errcode::{Kind, Origin};
+use ockam_transport_tcp::{Direction, PortalInterceptor};
 
 #[derive(Clone, Debug)]
 struct RequestInfo {
@@ -33,45 +33,98 @@ type CorrelationId = i32;
 pub(super) type TopicUuidMap = Arc<Mutex<HashMap<String, String>>>;
 
 #[async_trait]
-pub(crate) trait KafkaMessageInterceptor: Send + Sync + 'static {
+pub(crate) trait KafkaMessageInterceptorRequest: Send + Sync + 'static {
     async fn intercept_request(
         &self,
         context: &mut Context,
         original: BytesMut,
     ) -> Result<BytesMut, InterceptError>;
-
-    async fn intercept_response(
-        &self,
-        context: &mut Context,
-        original: BytesMut,
-    ) -> Result<BytesMut, InterceptError>;
-}
-
-#[derive(Clone)]
-pub(crate) struct InletInterceptorImpl {
-    request_map: Arc<Mutex<HashMap<CorrelationId, RequestInfo>>>,
-    uuid_to_name: TopicUuidMap,
-    secure_channel_controller: KafkaSecureChannelControllerImpl,
-    inlet_map: KafkaInletController,
-    encrypt_content: bool,
 }
 
 #[async_trait]
-impl KafkaMessageInterceptor for InletInterceptorImpl {
-    async fn intercept_request(
-        &self,
-        context: &mut Context,
-        original: BytesMut,
-    ) -> Result<BytesMut, InterceptError> {
-        self.intercept_request_impl(context, original).await
-    }
-
+pub(crate) trait KafkaMessageInterceptorResponse: Send + Sync + 'static {
     async fn intercept_response(
         &self,
         context: &mut Context,
         original: BytesMut,
-    ) -> Result<BytesMut, InterceptError> {
-        self.intercept_response_impl(context, original).await
+    ) -> Result<BytesMut, InterceptError>;
+}
+
+#[async_trait]
+pub(crate) trait KafkaMessageInterceptor:
+    KafkaMessageInterceptorRequest + KafkaMessageInterceptorResponse + Send + Sync + 'static
+{
+}
+
+pub struct KafkaMessageInterceptorWrapper {
+    decoder_from_inlet: Arc<Mutex<KafkaMessageDecoder>>,
+    decoder_from_outlet: Arc<Mutex<KafkaMessageDecoder>>,
+    message_interceptor: Arc<dyn KafkaMessageInterceptor>,
+    max_message_size: u32,
+}
+
+/// Converts a generic interceptor trait into kafka specific interceptor
+impl KafkaMessageInterceptorWrapper {
+    pub fn new(
+        message_interceptor: Arc<dyn KafkaMessageInterceptor>,
+        max_message_size: u32,
+    ) -> Self {
+        Self {
+            decoder_from_inlet: Arc::new(Mutex::new(KafkaMessageDecoder::new())),
+            decoder_from_outlet: Arc::new(Mutex::new(KafkaMessageDecoder::new())),
+            message_interceptor,
+            max_message_size,
+        }
+    }
+}
+
+#[async_trait]
+impl PortalInterceptor for KafkaMessageInterceptorWrapper {
+    async fn intercept(
+        &self,
+        context: &mut Context,
+        direction: Direction,
+        buffer: &[u8],
+    ) -> ockam_core::Result<Option<Vec<u8>>> {
+        let mut encoded_buffer: Option<BytesMut> = None;
+
+        let messages = {
+            let decoder = match direction {
+                Direction::FromOutletToInlet => &self.decoder_from_outlet,
+                Direction::FromInletToOutlet => &self.decoder_from_inlet,
+            };
+
+            let mut guard = decoder.lock().unwrap();
+            guard.extract_complete_messages(BytesMut::from(buffer), self.max_message_size)?
+        };
+
+        for complete_kafka_message in messages {
+            let transformed_message = match direction {
+                Direction::FromInletToOutlet => {
+                    self.message_interceptor
+                        .intercept_request(context, complete_kafka_message)
+                        .await
+                }
+                Direction::FromOutletToInlet => {
+                    self.message_interceptor
+                        .intercept_response(context, complete_kafka_message)
+                        .await
+                }
+            }
+            .map_err(|error| match error {
+                InterceptError::Io(cause) => Error::new(Origin::Transport, Kind::Io, cause),
+                InterceptError::Ockam(error) => error,
+            })?;
+
+            // avoid copying the first message
+            if let Some(encoded_buffer) = encoded_buffer.as_mut() {
+                encoded_buffer.extend_from_slice(length_encode(transformed_message)?.as_ref());
+            } else {
+                encoded_buffer = Some(length_encode(transformed_message)?)
+            }
+        }
+
+        Ok(encoded_buffer.map(|buffer| buffer.freeze().to_vec()))
     }
 }
 
@@ -83,19 +136,12 @@ struct MessageWrapper {
     #[n(1)] content: Vec<u8>
 }
 
-impl InletInterceptorImpl {
-    pub(crate) fn new(
-        secure_channel_controller: KafkaSecureChannelControllerImpl,
-        uuid_to_name: TopicUuidMap,
-        inlet_map: KafkaInletController,
-        encrypt_content: bool,
-    ) -> InletInterceptorImpl {
-        Self {
-            request_map: Arc::new(Mutex::new(Default::default())),
-            uuid_to_name,
-            secure_channel_controller,
-            inlet_map,
-            encrypt_content,
-        }
-    }
+/// By default, kafka supports up to 1MB messages. 16MB is the maximum suggested
+pub(crate) const MAX_KAFKA_MESSAGE_SIZE: u32 = 16 * 1024 * 1024;
+
+// internal error to return both io and ockam errors
+#[derive(Debug)]
+pub(crate) enum InterceptError {
+    Io(ockam_core::compat::io::Error),
+    Ockam(ockam_core::Error),
 }
