+++ implementations/rust/ockam/ockam_api/src/kafka/tests/mod.rs
@@ -1,0 +1,2 @@
+mod integration_test;
+mod interceptor_test;
