+++ implementations/rust/ockam/ockam_api/src/nodes/service/kafka_services.rs
@@ -1,21 +1,12 @@
-use ockam::transport::HostnamePort;
-use ockam::{Address, Context, Result};
-use ockam_abac::PolicyExpression;
-use ockam_abac::{Action, Resource, ResourceType};
-use ockam_core::api::{Error, Response};
-use ockam_core::compat::rand::random_string;
-use ockam_core::route;
-use ockam_multiaddr::proto::Project;
-use ockam_multiaddr::MultiAddr;
-use std::sync::Arc;
-
 use super::NodeManagerWorker;
 use crate::error::ApiError;
-use crate::kafka::secure_channel_map::controller::KafkaSecureChannelControllerImpl;
-use crate::kafka::OutletManagerService;
+use crate::kafka::key_exchange::controller::KafkaKeyExchangeController;
+use crate::kafka::protocol_aware::inlet::KafkaInletInterceptorFactory;
+use crate::kafka::protocol_aware::outlet::KafkaOutletInterceptorFactory;
+use crate::kafka::KafkaOutletController;
 use crate::kafka::{
     kafka_policy_expression, ConsumerPublishing, ConsumerResolution, KafkaInletController,
-    KafkaPortalListener, KAFKA_OUTLET_BOOTSTRAP_ADDRESS, KAFKA_OUTLET_INTERCEPTOR_ADDRESS,
+    KAFKA_OUTLET_BOOTSTRAP_ADDRESS, KAFKA_OUTLET_INTERCEPTOR_ADDRESS,
 };
 use crate::nodes::models::portal::OutletAccessControl;
 use crate::nodes::models::services::{
@@ -25,6 +16,18 @@ use crate::nodes::registry::{KafkaServiceInfo, KafkaServiceKind};
 use crate::nodes::service::default_address::DefaultAddress;
 use crate::nodes::InMemoryNode;
 use crate::port_range::PortRange;
+use ockam::transport::HostnamePort;
+use ockam::{Address, Context, Result};
+use ockam_abac::PolicyExpression;
+use ockam_abac::{Action, Resource, ResourceType};
+use ockam_core::api::{Error, Response};
+use ockam_core::compat::rand::random_string;
+use ockam_core::flow_control::FlowControls;
+use ockam_core::route;
+use ockam_multiaddr::proto::Project;
+use ockam_multiaddr::MultiAddr;
+use ockam_transport_tcp::{PortalInletInterceptor, PortalOutletInterceptor};
+use std::sync::Arc;
 
 impl NodeManagerWorker {
     pub(super) async fn start_kafka_inlet_service(
@@ -109,7 +112,7 @@ impl InMemoryNode {
     pub async fn start_kafka_inlet_service(
         &self,
         context: &Context,
-        local_interceptor_address: Address,
+        interceptor_address: Address,
         bind_address: HostnamePort,
         brokers_port_range: (u16, u16),
         outlet_node_multiaddr: MultiAddr,
@@ -124,7 +127,7 @@ impl InMemoryNode {
             .policy_access_control(
                 self.project_authority().clone(),
                 Resource::new(
-                    format!("kafka-consumer-{}", local_interceptor_address.address()),
+                    format!("kafka-consumer-{}", interceptor_address.address()),
                     ResourceType::KafkaConsumer,
                 ),
                 Action::HandleMessage,
@@ -136,7 +139,7 @@ impl InMemoryNode {
             .policy_access_control(
                 self.project_authority().clone(),
                 Resource::new(
-                    format!("kafka-producer-{}", local_interceptor_address.address()),
+                    format!("kafka-producer-{}", interceptor_address.address()),
                     ResourceType::KafkaProducer,
                 ),
                 Action::HandleMessage,
@@ -144,7 +147,7 @@ impl InMemoryNode {
             )
             .await?;
 
-        let secure_channel_controller = KafkaSecureChannelControllerImpl::new(
+        let secure_channel_controller = KafkaKeyExchangeController::new(
             self.node_manager.clone(),
             self.secure_channels.clone(),
             consumer_resolution,
@@ -173,8 +176,9 @@ impl InMemoryNode {
         };
 
         let inlet_controller = KafkaInletController::new(
+            self.node_manager.clone(),
             outlet_node_multiaddr.clone(),
-            route![local_interceptor_address.clone()],
+            route![interceptor_address.clone()],
             route![KAFKA_OUTLET_INTERCEPTOR_ADDRESS],
             bind_address.hostname(),
             PortRange::try_from(brokers_port_range)
@@ -196,7 +200,7 @@ impl InMemoryNode {
         self.create_inlet(
             context,
             bind_address,
-            route![local_interceptor_address.clone()],
+            route![interceptor_address.clone()],
             route![
                 KAFKA_OUTLET_INTERCEPTOR_ADDRESS,
                 KAFKA_OUTLET_BOOTSTRAP_ADDRESS
@@ -216,21 +220,20 @@ impl InMemoryNode {
         let policy_access_control = self
             .policy_access_control(
                 self.project_authority().clone(),
-                Resource::new(
-                    local_interceptor_address.to_string(),
-                    ResourceType::TcpInlet,
-                ),
+                Resource::new(interceptor_address.to_string(), ResourceType::TcpInlet),
                 Action::HandleMessage,
                 inlet_policy_expression,
             )
             .await?;
 
-        KafkaPortalListener::create(
+        PortalInletInterceptor::create(
             context,
-            encrypt_content,
-            inlet_controller,
-            secure_channel_controller,
-            local_interceptor_address.clone(),
+            interceptor_address.clone(),
+            Arc::new(KafkaInletInterceptorFactory::new(
+                secure_channel_controller,
+                inlet_controller,
+                encrypt_content,
+            )),
             Arc::new(policy_access_control.create_incoming()),
             Arc::new(policy_access_control.create_outgoing(context).await?),
         )
@@ -239,7 +242,7 @@ impl InMemoryNode {
         self.registry
             .kafka_services
             .insert(
-                local_interceptor_address,
+                interceptor_address,
                 KafkaServiceInfo::new(KafkaServiceKind::Inlet),
             )
             .await;
@@ -271,39 +274,58 @@ impl InMemoryNode {
             )
             .await?;
 
-        OutletManagerService::create(
-            context,
-            default_secure_channel_listener_flow_control_id,
+        let spawner_flow_control_id = FlowControls::generate_flow_control_id();
+        let outlet_controller = KafkaOutletController::new(
+            self.node_manager.clone(),
             outlet_policy_expression.clone(),
-            Arc::new(policy_access_control.create_incoming()),
-            Arc::new(policy_access_control.create_outgoing(context).await?),
             tls,
+        );
+        let interceptor_address = Address::from_string(KAFKA_OUTLET_INTERCEPTOR_ADDRESS);
+
+        PortalOutletInterceptor::create(
+            context,
+            interceptor_address.clone(),
+            Some(spawner_flow_control_id.clone()),
+            Arc::new(KafkaOutletInterceptorFactory::new(
+                outlet_controller.clone(),
+                spawner_flow_control_id.clone(),
+            )),
+            Arc::new(policy_access_control.create_outgoing(context).await?),
+            Arc::new(policy_access_control.create_incoming()),
         )
         .await?;
 
-        if let Err(e) = self
-            .create_outlet(
-                context,
-                bootstrap_server_addr,
-                tls,
-                Some(KAFKA_OUTLET_BOOTSTRAP_ADDRESS.into()),
-                false,
-                OutletAccessControl::WithPolicyExpression(outlet_policy_expression),
-            )
-            .await
-        {
-            return Err(ApiError::core(e.to_string()));
-        };
+        // every secure channel can reach this service
+        let flow_controls = context.flow_controls();
+        flow_controls.add_consumer(
+            interceptor_address.clone(),
+            &default_secure_channel_listener_flow_control_id,
+        );
 
-        {
-            self.registry
-                .kafka_services
-                .insert(
-                    service_address,
-                    KafkaServiceInfo::new(KafkaServiceKind::Outlet),
-                )
-                .await;
-        }
+        // this spawner flow control id is used to control communication with dynamically created
+        // outlets
+        flow_controls.add_spawner(interceptor_address.clone(), &spawner_flow_control_id);
+
+        // allow communication with the kafka bootstrap outlet
+        flow_controls.add_consumer(KAFKA_OUTLET_BOOTSTRAP_ADDRESS, &spawner_flow_control_id);
+
+        self.create_outlet(
+            context,
+            bootstrap_server_addr,
+            tls,
+            Some(KAFKA_OUTLET_BOOTSTRAP_ADDRESS.into()),
+            false,
+            OutletAccessControl::WithPolicyExpression(outlet_policy_expression),
+        )
+        .await?;
+
+        self.registry
+            .kafka_services
+            .insert(
+                service_address,
+                KafkaServiceInfo::new(KafkaServiceKind::Outlet),
+            )
+            .await;
 
         Ok(())
     }
