+++ implementations/rust/ockam/ockam_transport_tcp/Cargo.toml
@@ -32,6 +32,7 @@ ring = ["tokio-rustls/ring"]
 
 [dependencies]
 cfg-if = "1.0.0"
+log = "0.4"
 minicbor = "0.24"
 ockam_core = { path = "../ockam_core", version = "^0.114.0" }
 ockam_macros = { path = "../ockam_macros", version = "^0.34.0" }
