+++ implementations/rust/ockam/ockam_transport_tcp/src/portal/mod.rs
@@ -1,5 +1,6 @@
 mod addresses;
 mod inlet_listener;
+mod interceptor;
 pub mod options;
 mod outlet_listener;
 mod portal_message;
@@ -7,6 +8,10 @@ mod portal_receiver;
 mod portal_worker;
 
 pub(crate) use inlet_listener::*;
+pub use interceptor::{
+    Direction, PortalInletInterceptor, PortalInterceptor, PortalInterceptorFactory,
+    PortalInterceptorWorker, PortalOutletInterceptor,
+};
 pub(crate) use outlet_listener::*;
 pub use portal_message::*;
 pub(crate) use portal_receiver::*;
